Refactored jenkins pipeline
